add view_date to get planning distributions (#534)
Allows the retrieval of planned backlog for a past date.

Adds view_date parameter to /planning_distributions endpoint, if this parameter is present all calculations will be made as if the current date were view_date.

To implement this behavior, the way in which forecasts and deviations are read from the database is modified (this modifications are only executed if the view_date is parameter is present, if it isn't then the flow of execution is as before).

The rows from planning_distribution that will be read are those that belong to the last forecast that was uploaded before view_date, that is, the one with forecast.date_created <= view_date.

For deviations, it must be noted, that at any time at most one deviation can be active. So for a given view_date:

There is no deviation for that warehouse, so no deviation is applied.
There were no active deviations at the time of view_date, all deviations satisfy: date_created < view_date ---> last_updated < view_date.
There is one and only deviation such that: date_created <= view_date ---> view_date < last_updated.
The query findActiveDeviationAt will work only as these conditions are met.